ml api 최종

+++ ml_api/src/mlagent.py
@@ -12,34 +12,8 @@ MLAgent = Namespace(
 env_path = "/mlagent/Project-MetaDogs/Linux-build/AITest.x86_64"
 model_name = "AITest"
 model_path = f"/mlagent/{model_name}.yaml"
-result_path = "/mlagent/usr_model"
+result_path = "/mlagent/pet_model"
 #오닉스 이름: yaml파일과 동일
-
-@MLAgent.route('/create_model')
-class Create_Model(Resource):
-    def post(self):
-        # Get data
-        data = request.get_json()
-
-        run_id = f"{result_path}/{data['pet_token']}/model_{data['gesture_id']}"
-
-        #토큰으로 폴더 생성
-        cmd = f"mkdir {result_path}/{data['pet_token']}"
-        system_return = os.system(cmd)
-
-        #첫 학습
-        cmd = f"mlagents-learn {model_path} --run-id={run_id} --env={env_path} --no-graphics"
-        system_return = os.system(cmd)
-
-        #오닉스 db에 저장
-        SyncModelToDB(data['pet_token'], data['gesture_id'], f"{run_id}/{model_name}.onnx")
-
-        #오닉스 파일 리턴
-        try:
-             return send_file(f"{run_id}/{model_name}.onnx", attachment_filename=f"model_{data['gesture_id']}")
-        
-        except Exception as e:
-            return str(e)
 
 @MLAgent.route('/learning')
 class Learning(Resource):
@@ -47,27 +21,41 @@ class Learning(Resource):
         # Get data
         data = request.get_json()
 
-        old_run_id = f"{result_path}/{data['pet_token']}/model_{data['gesture_id']}"
+        pet_token_path = f"mkdir {result_path}/{data['pet_token']}"
+        old_run_id = f"{pet_token_path}/model_{data['gesture_id']}"
         new_run_id = old_run_id + "_new" #새 오닉스 파일이 저장될 폴더
+        
 
-        #재학습
-        cmd = f"mlagents-learn {model_path}  --initialize-from={old_run_id} --run-id={new_run_id} --env={env_path} --no-graphics"
-        system_return = os.system(cmd) #0 리턴: 성공
+        if not os.path.isdir(pet_token_path):
+            #토큰으로 폴더 생성
+            cmd = f"mkdir {pet_token_path}"
+            system_return = os.system(cmd)  #0 리턴하면 성공
 
-        #기존 모델 삭제
-        cmd = f"rm -rf {old_run_id}"
-        system_return = os.system(cmd)
+        if not os.path.isdir(old_run_id):
+            #첫 학습
+            cmd = f"mlagents-learn {model_path} --run-id={old_run_id} --env={env_path} --no-graphics"
+            system_return = os.system(cmd)
 
-        #새 모델 이름 변경
-        cmd = f"mv {new_run_id} {old_run_id}"
-        system_return = os.system(cmd)
+        else:
+            #재학습
+            cmd = f"mlagents-learn {model_path}  --initialize-from={old_run_id} --run-id={new_run_id} --env={env_path} --no-graphics"
+            system_return = os.system(cmd)
+
+            #기존 모델 삭제
+            cmd = f"rm -rf {old_run_id}"
+            system_return = os.system(cmd)
+
+            #새 모델 이름 변경
+            cmd = f"mv {new_run_id} {old_run_id}"
+            system_return = os.system(cmd)
+        
 
         #오닉스 db에 저장
         SyncModelToDB(data['pet_token'], data['gesture_id'], f"{old_run_id}/{model_name}.onnx")
 
         #오닉스 파일 리턴
         try:
-             return send_file(f"{old_run_id}/{model_name}.onnx", attachment_filename=f"model_{data['gesture_id']}")
+             return send_file(f"{old_run_id}/{model_name}.onnx", download_name=f"model_{data['gesture_id']}.onnx")
         
         except Exception as e:
             return str(e)
